+++ docs/manual/po4_1.py
@@ -1,12 +1,13 @@
 # import classes from the esbmtk library
 from esbmtk import (
     Model,  # the model class
-    Species,  # the reservoir class
-    Connect,  # the connection class
-    Source,  # the source class
-    Sink,  # sink class
+    Reservoir,  # the reservoir class
+    ConnectionProperties,  # the connection class
+    SourceProperties,  # the source class
+    SinkProperties,  # sink class
     Q_,  # Quantity operator
 )
+# Note that complete code examples are available from
 
 # define the basic model parameters
 M = Model(
@@ -22,34 +23,32 @@ F_b = 0.01  # About 1% of the exported P is buried in the deep ocean
 thc = "20*Sv"  # Thermohaline circulation in Sverdrup
 
 # Source definitions
-Source(
+SourceProperties(
     name="weathering",
-    species=M.PO4,
+    species=[M.PO4],
     register=M,  # i.e., the instance will be available as M.weathering
 )
-Sink(
+SinkProperties(
     name="burial",
-    species=M.PO4,
+    species=[M.PO4],
     register=M,  #
 )
 
 # reservoir definitions
-Species(
+Reservoir(
     name="S_b",  # box name
-    species=M.PO4,  # species in box
     register=M,  # this box will be available as M.S_b
     volume="3E16 m**3",  # surface box volume
-    concentration="0 umol/l",  # initial concentration
+    concentration={M.PO4: "0 umol/l"},  # initial concentration
 )
-Species(
+Reservoir(
     name="D_b",  # box name
-    species=M.PO4,  # species in box
     register=M,  # this box will be available M.D_b
     volume="100E16 m**3",  # deeb box volume
-    concentration="0 umol/l",  # initial concentration
+    concentration={M.PO4: "0 umol/l"},  # initial concentration
 )
 
-Connect(
+ConnectionProperties(
     source=M.weathering,  # source of flux
     sink=M.S_b,  # target of flux
     rate=F_w,  # rate of flux
@@ -57,7 +56,7 @@ Connect(
     ctype="regular",
 )
 
-Connect(  # thermohaline downwelling
+ConnectionProperties(  # thermohaline downwelling
     source=M.S_b,  # source of flux
     sink=M.D_b,  # target of flux
     ctype="scale_with_concentration",
@@ -65,7 +64,7 @@ Connect(  # thermohaline downwelling
     id="downwelling_PO4",
     # ref_reservoirs=M.S_b, defaults to the source instance
 )
-Connect(  # thermohaline upwelling
+ConnectionProperties(  # thermohaline upwelling
     source=M.D_b,  # source of flux
     sink=M.S_b,  # target of flux
     ctype="scale_with_concentration",
@@ -73,23 +72,25 @@ Connect(  # thermohaline upwelling
     id="upwelling_PO4",
 )
 
-Connect(  #
+ConnectionProperties(  #
     source=M.S_b,  # source of flux
     sink=M.D_b,  # target of flux
     ctype="scale_with_concentration",
     scale=M.S_b.volume / tau,
     id="primary_production",
+    species=[M.PO4],  # apply this only to PO4
 )
 
-Connect(  #
+ConnectionProperties(  #
     source=M.D_b,  # source of flux
     sink=M.burial,  # target of flux
     ctype="scale_with_flux",
     ref_flux=M.flux_summary(filter_by="primary_production", return_list=True)[0],
     scale=F_b,
     id="burial",
+    species=[M.PO4],
 )
 
 M.run()
-M.plot([M.S_b, M.D_b])
+M.plot([M.S_b.PO4, M.D_b.PO4])
 M.save_data()